+++ cpp/src/dbscan/vertexdeg/runner.h
@@ -25,9 +25,9 @@ namespace Dbscan {
 namespace VertexDeg {
 
 template <typename Type_f, typename Index_ = long>
-void run(const ML::cumlHandle_impl& handle, bool* adj, int* vd, Type_f* x,
-         Type_f eps, Index_ N, Index_ D, int algo, Index_ startVertexId,
-         Index_ batchSize, cudaStream_t stream) {
+void run(const ML::cumlHandle_impl& handle, bool* adj, unsigned long long* vd,
+         Type_f* x, Type_f eps, Index_ N, Index_ D, int algo,
+         Index_ startVertexId, Index_ batchSize, cudaStream_t stream) {
   Pack<Type_f, Index_> data = {vd, adj, x, eps, N, D};
   switch (algo) {
     case 0:
