+++ cpp/src/metrics/metrics.hpp
@@ -106,10 +106,10 @@ double silhouetteScore(const cumlHandle &handle, double *y, int nRows,
 * @param upper_class_range: the highest value in the range of classes
 * @return: The adjusted rand index value
 */
-double adjustedRandIndex(const cumlHandle &handle, const int *y,
-                         const int *y_hat, const int n,
-                         const int lower_class_range,
-                         const int upper_class_range);
+double adjustedRandIndex(const cumlHandle &handle, const long *y,
+                         const long *y_hat, const long n,
+                         const long lower_class_range,
+                         const long upper_class_range);
 
 /**
 * Calculates the "Kullback-Leibler Divergence"
